[2/n] Update Token registration and discarding on sign out (#3)
> [!NOTE]
> This PR is based on other PRs and needs those to be merged before this
one can get merged:
> - #4 

This PR adjusts the current APNS implementation:
- support for the new APNS payload format
- support for discarding tokens on logout

For this _signed in_ state (in the form of credentials) had to be passed
into the `useToken` hook, so it can react to signing out (and back in
again).

Also this PR cleans up unused dependencies such as `react-native-auth0`
and `aws-amplify`. Let me know if that makes it harder to review and I
gladly split this part into a separate PR.


## Test Plan

Tested in iOS Smulator. Push tokens are registered on sign in, and
discarded on signet. Push notifications arrive and work as expected.