[3/n] Migrating to `expo-notification` and adding support for Android (#2)
> [!NOTE]
> This PR is based on other PRs and needs those to be merged before this
one can get merged:
> - #4 
> - #3 

This PR is mostly about migrating to
[expo-notifications](https://docs.expo.dev/versions/latest/sdk/notifications/)
which supports push notifications for both iOS and Android (as well as
web push as well, if that's something we want to explore in the future).

This updates a lot of the native code, but brings the benefit that most
of the custom added native code becomes unnecessary and is handled by
expo behind the scenes.

So this PR does the remaining work of adding Android and FCM support. It
introduces platform splitting around push token registration and push
notification handling.

As a small plus of using Expo we now can easily deduce the value for
`installationID` (and therefore which APNS environment will be used)
from the native iOS side.

## TestPlan

In order to test push tokens you need to have a firebase app registered
that corresponds to the bundleID (`com.rnprototype`). You'll need to put
the according `google-services.json` under
`android/app/google-services.json`.

Also the FCM integration needs to be setup. (let me know if you need me
to provide steps).

Tested both iOS and Android in sim-/emulator. The apps register and
discard tokens correctly on signing and signout. Push notifications
arrive on both platforms and can be opened.