chore: update reanimated to fix android build (#9)
## Change description

The Android build was brooken because our version of reanimated was
incompatible with the react-native version.

Compatibility matrix:
https://docs.swmansion.com/react-native-reanimated/docs/guides/compatibility

We were on `3.7.2` which was only compatible until react-native `0.73`

## Test Plan

`yarn start` and pressing `a` to build and run the android build. It
failed before the change, and works now.
Also tested iOS build by pressing `i`. Also still works.

## Type of Change

- [ ] Bug fix      <!-- fixes an issue -->
- [ ] Feature      <!-- adds functionality -->
- [x] Enhancement  <!-- improves something existing -->