Moving to Expo & Continuous Native Generation (#20)
## Change description

This is a big change to adopt [Continuous Native
Generation](https://docs.expo.dev/workflow/continuous-native-generation/).
The gist is, instead of checking in the native code in `/android` and
`/ios` CNG propagates those folders from a template, together with the
info provided in `app.json` and
[config-plugins](https://docs.expo.dev/config-plugins/introduction/)
that are part of node-modules.

This is different than using Expo GO, which is a prebuild native
application that just loads the JavaScript parts of the app. With CNG we
can still build a the native apps locally, and link native modules.

Parts of this PR:
- Removes native code and replaces it with `app.json` and `npx expo
prebuild`.
- Update yarn scripts to automate prebuild (i.e. `yarn ios` and `yarn
android` calls it automatically)
- Replaced font loading and assets. Introduced splash screen component
during font loading.
- Commited `google-services.json` file ("contains unique but non-secret
identifiers" and can be checked in
([source](https://firebase.google.com/docs/projects/learn-more#config-files-objects)))

## Test Plan

<!-- How did you test the changes, what are the suggested steps for
reviewers if they want to test your changes? -->

## Type of Change

- [ ] Bug fix      <!-- fixes an issue -->
- [ ] Feature      <!-- adds functionality -->
- [ ] Enhancement  <!-- improves something existing -->